refactor[api-server]: lottery.error 패키지 위치 변경
