+++ src/main/java/com/watermelon/server/event/lottery/error/ExpectationAlreadyExistError.java
@@ -1,4 +1,4 @@
-package com.watermelon.server.lottery.error;
+package com.watermelon.server.event.lottery.error;
 
 public class ExpectationAlreadyExistError extends Exception{
 }
