+++ src/main/java/com/watermelon/server/event/lottery/error/ExpectationNotExist.java
@@ -1,4 +1,4 @@
-package com.watermelon.server.lottery.error;
+package com.watermelon.server.event.lottery.error;
