+++ src/main/java/com/watermelon/server/lottery/service/ExpectationService.java
@@ -5,8 +5,8 @@ import com.watermelon.server.event.lottery.domain.Expectation;
 import com.watermelon.server.event.lottery.domain.LotteryApplier;
 import com.watermelon.server.event.lottery.dto.request.RequestExpectationDto;
 import com.watermelon.server.event.lottery.dto.response.ResponseExpectationDto;
-import com.watermelon.server.lottery.error.ExpectationAlreadyExistError;
-import com.watermelon.server.lottery.error.ExpectationNotExist;
+import com.watermelon.server.event.lottery.error.ExpectationAlreadyExistError;
+import com.watermelon.server.event.lottery.error.ExpectationNotExist;
 import com.watermelon.server.lottery.repository.ExpectationRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
